explorationEditor utility
