+++ core/tests/puppeteer-acceptance-tests/puppeteer-testing-utilities/user-factory.ts
@@ -29,10 +29,8 @@ import {
 import {BlogAdminFactory, BlogAdmin} from '../user-utilities/blog-admin-utils';
 import {QuestionAdminFactory} from '../user-utilities/question-admin-utils';
 import {BlogPostEditorFactory} from '../user-utilities/blog-post-editor-utils';
-import {
-  VoiceoverAdminFactory,
-  VoiceoverAdmin,
-} from '../user-utilities/voiceover-admin-utils';
+import {VoiceoverAdminFactory} from '../user-utilities/voiceover-admin-utils';
+import {ExplorationEditorFactory} from '../user-utilities/exploration-creator-utils';
 import testConstants from './test-constants';
 
 const ROLES = testConstants.Roles;
@@ -143,13 +141,11 @@ export class UserFactory {
     username: string,
     email: string,
     roles: OptionalRoles<TRoles> = [] as OptionalRoles<TRoles>
-  ): Promise<LoggedInUser & VoiceoverAdmin & MultipleRoleIntersection<TRoles>> {
-    let user = UserFactory.composeUserWithRoles(
-      UserFactory.composeUserWithRoles(BaseUserFactory(), [
-        LoggedInUserFactory(),
-      ]),
-      [VoiceoverAdminFactory()]
-    );
+  ): Promise<LoggedInUser & MultipleRoleIntersection<TRoles>> {
+    let user = UserFactory.composeUserWithRoles(BaseUserFactory(), [
+      LoggedInUserFactory(),
+      ExplorationEditorFactory(),
+    ]);
     await user.openBrowser();
     await user.signUpNewUser(username, email);
     activeUsers.push(user);
